pyproject.toml
[tool.poetry]
name = "beam-python-poc"
version = "0.1.0"
description = ""
authors = ["David <davigetto@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
confluent-kafka = "^1.8.2"
kafka-python = "^2.0.2"
apache-beam = "^2.37.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
